requirements.txt
PyQt5==5.15.9
openai-whisper==20230918
sounddevice==0.4.6
soundfile==0.12.1
pynput==1.7.6
pyperclip==1.8.2
numpy==1.24.3
Pillow==10.0.0
imageio-ffmpeg==0.6.0
ffmpeg-python==0.2.0

